infrastructure/requirements-dev.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pyyaml>=6.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
coverage>=7.0.0